Varios
- alterada camera para a camera natica iOS
- camera agora deixa editar as fotos
- o picker da quantidade dos ingredientes foi dividido em 2 ( textview
e picker)
- o picker dos ingredientes dentro do carrinho foi alterado para ficar
igual ao da receita
- primeiro caracter nome da receita agora começa capitalizado em criar
receitas
- primeiro caracter nome de Ingrediente agora começa capitalizado em
criar ingredientes
- primeiro caracter nome de Etapa agora começa capitalizado em criar
etapas
- primero caracter de notas ao criar notas agora também começa
capitalizado

+++ Fucook/ListaCompras/ListaCompras.h
@@ -10,7 +10,7 @@
 
 @interface ListaCompras : UIViewController <UITableViewDelegate, UITableViewDataSource, UIPickerViewDataSource, UIPickerViewDelegate> {
     
-    int selectedIndex;    
+    NSInteger selectedIndex;
     NSMutableArray * arrayOfItems;
     
 }
